scm_p_require: add missing GC protection for a local variable (#21)
Fixes #20.

Original code, when SCM_COMPAT_SIOD was enabled, allowed the `filename`
object to be freed by GC during the call to `scm_require_internal`,
causing a use-after-free problem. This was caused by the C compiler not
saving the `filename` pointer on the stack across the call.

This patch fixes it by marking `filename` volatile. (I can think of many
ways to fix the problem. I chose `volatile` because it's simple and it's
already used elsewhere in the codebase)