Add GCD and LCM
The algorithm used in GCD is Euclid's. Since R5RS requires it for
several arguments, and Sigscheme implements other variable-length
procedures as folds, this implementation follows that. Basically,

(gcd a1 a2 a3 ...)

is implemented as

(fold 0 %gcd2 '(a1 a2 a3 ...))

Here, "fold" does not exist - it's written in C, in a similar
fashion to other functions (+, /, * etc), and "%gcd2" is a
static C function.

LCM is implemented as usual using the GCD.